Finish Table resize
