+++ src/components/table/table.function.js
@@ -1,0 +1,3 @@
+export function shouldResize(event) {
+  return event.target.dataset.resize;
+}
